Melhora no historico e correções geral
Feito melhorias no codigo da pagina do historico além de correções em outras paginas conforme foi solicitado
