+++ planit/app/(tabs)/Home.tsx
@@ -209,10 +209,13 @@ const Home = () => {
     buscarAgendamentos();
   }, [buscarAgendamentos]);
 
-  // Função auxiliar para formatar a data
   const formatDate = (date: Date) => {
-    return date.toISOString().split('T')[0]; // Formato YYYY-MM-DD
-  };
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0'); 
+  const day = String(date.getDate()).padStart(2, '0');
+  
+  return `${year}-${month}-${day}`; 
+};
 
   // Carrega as consultas da semana (mantém a lógica original)
   const carregarConsultasSemana = useCallback(async () => {
